+++ src/dpr/components/_inputs/autocomplete-text-input/clientClass.ts
@@ -37,6 +37,7 @@ class Autocomplete extends DprClientClass {
 
       if (hiddenInput) {
         hiddenInput.value = ''
+        hiddenInput.disabled = true
       }
 
       delete textInput.dataset.staticOptionNameValue
@@ -57,9 +58,17 @@ class Autocomplete extends DprClientClass {
 
   initialiseDefaultValue(textInput) {
     const hiddenInput = this.getHiddenInput()
-
     if (hiddenInput?.value) {
-      textInput.dataset.staticOptionNameValue = hiddenInput.value
+      hiddenInput.disabled = false
+      return
+    }
+
+    textInput.value = ''
+    delete textInput.dataset.staticOptionNameValue
+
+    if (hiddenInput) {
+      hiddenInput.value = ''
+      hiddenInput.disabled = true
     }
   }
 
@@ -151,6 +160,7 @@ class Autocomplete extends DprClientClass {
     // submission value
     if (hiddenInput) {
       hiddenInput.value = actualValue
+      hiddenInput.disabled = false
     }
 
     textInput.dataset.staticOptionNameValue = actualValue